Added categories for reviews in the API. Changed the API so that the associated Application and Category objects are always returned within the Review object in responses. Fixed model associations.

+++ API/MobileAPI/server/app.js
@@ -23,17 +23,17 @@ dbSync = async () => {
         });
 
     //In development, force tables to the db
-    /*if(process.env.NODE_ENV == 'dev' || process.env.NODE_ENV == 'test') {
+    if(/*process.env.NODE_ENV == 'dev' ||*/ process.env.NODE_ENV == 'test') {
         await models.sequelize.sync({force: true})
             .then(() => {
                 console.log('Models loaded and forcefully synced.');
             });
-    } else {*/
+    } else {
         await models.sequelize.sync()
             .then(() => {
                 console.log('Models loaded and synced.');
             });
-    //}
+    }
     await syncCategories();
     console.log('Categories synced from JSON.');
 };
